restructure and add babel build
