+++ .gitignore
@@ -1,3 +1,6 @@
+#build
+lib/*
+
 # Logs
 logs
 *.log
