--- index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  createPlugin: require('./src/plugin'),
-  shouldUpdateReducer: require('./src/reducer'),
-}
