+++ package.json
@@ -1,11 +1,13 @@
 {
   "name": "@synapsestudios/fetch-client-goalie",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Fetch client plugin and redux bindings for dealing with goalie 412 responses",
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "lint": "eslint ./src"
+    "lint": "eslint ./src",
+    "clean": "rm -rf lib",
+    "build": "npm run clean && babel src --out-dir lib"
   },
   "repository": {
     "type": "git",
@@ -18,6 +20,7 @@
   },
   "homepage": "https://github.com/synapsestudios/fetch-client-goalie#readme",
   "devDependencies": {
+    "babel-cli": "^6.26.0",
     "eslint": "^4.19.1",
     "eslint-plugin-jest": "^21.17.0",
     "jest": "^23.0.1"
