+++ src/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  createPlugin: require('./plugin'),
+  shouldUpdateReducer: require('./reducer'),
+}
